Changes:
- move field tests to fields
- remove duplicate test file in run_sync
- add duration field test
